Fix Huobi's POST requests (#186)
* Fix Huobi's POST request content-type

If we try to send a POST request to Huobi w/ a content-type of
"application/x-www-form-urlencoded" we will receive an error:

Content type 'application/x-www-form-urlencoded;charset=UTF-8' not supported

Their english documentation is incorrect, it does not specify this atm,
translating their chinese documentation, it states:

"Content-Type: application/json must be declared in the POST request header;
Content-Type: application/x-www-form-urlencoded must be declared in the GET
request header. (Chinese users recommend setting Accept-Language: zh-cn)"

* Fix Huobi's place new order request (send as json)

We should not send the order details through url parameters, this
needs to be sent as a json payload via the request body.

Documentation:

https://github.com/huobiapi/API_Docs_en/wiki/REST_Reference#post-v1orderordersplace--make-an-order-in-huobipro

* Fix Huobi's margin transfer/margin order/withdraw requests

This requests data need to be sent as json, not as a query string.

Docs:

https://github.com/huobiapi/API_Docs_en/wiki/REST_Reference#post-v1dwtransfer-inmargin--transfer-asset-from-spot-account-to-margin-account

https://github.com/huobiapi/API_Docs_en/wiki/REST_Reference#post-v1marginorders--margin-application

https://github.com/huobiapi/API_Docs_en/wiki/REST_Reference#post-v1dwwithdrawapicreate---create-a-withdraw-application

* Fix Huobi's margin repayment request

This request data needs to be sent a json, not via the query string.

Also note, that the "order-id" parameter is already sent through the
url path, it does not need to be included in the request body.