Add support for multi-instance exchanges
This allows you to have multiple instances of the same exchange that
might be configured differently. (eg. two instances of Binance with
different credentials). This functionality is also a preparation and
paves the way for multiple instances of in-house custom exchanges
that might be used for load/correctness/consistency/validation testing.