Coinut: Fix WS not parsing instruments and err
WsConnect is calling GetInstruments, and when that fails, erroring out
and not subscribing to anything.

The response to get instruments is an object {}, which skips the
IncomingWithData check in wsReadData for arrays [].
The check in wsHandleData depended on client_ord_id, but I figure so
long as there's a nonce which matches, we can palm it off.

This results simutaneously in having to move the login handler back to
it's waiting nonce-parser, and also simplifying/deduping it.