Huobi: Fix TestPairFromContractExpiryCode intermitent failure
Huobi decided to not even return CW at all in [this
test](https://github.com/thrasher-corp/gocryptotrader/actions/runs/17637897980/job/50117883059?pr=1990#step:11:1280)

Reverse the test expectation. Just get *something* back in contract
codes. We have to accept that runtime won't work sometimes when we want
to convert codes that they haven't told us about... and that's the way
it is.