Deribit: Fix future_combo instrument parsing
Deribit seems to have opened up more instruments for future combos and
is using a `-` delimiter in them, throwing off our parsing rule that
it's always 3 parts for them.

This change adds in the "FS" part, which seems consistent.

I would have added checking for such a part to option_combo but some of
them are missing it :/